code cleanup, delete questions page replace with stepper
